// src/app/claim-processing/manual-claim-processing/process-claim-medical-visit/model/proc-clm-med-visit-chrg.model.ts
import {ScreenChargeLine} from './screen-charge-line.model';

/**
 * Model class ProcClmMedVisitChrg
 * Path: screenbean/procclmmedvisitchrg
 * Model: com::uhc::aarp::fox::domain::screenbean::procclmmedvisitchrg::ProcClmMedVisitChrg
 */
export class ProcClmMedVisitChrg {
  m25cmnd = '';
  m25mem1 = '';
  m25mem2 = '';
  m25mem3 = '';
  m25cno1 = '';
  m25cno2 = '';
  m25cno3 = '';
  m25cno4 = '';
  m25cno5 = '';
  m25cno6 = '';
  m25com1 = '';
  m25com2 = '';
  m25name = '';
  m25pln1 = '';
  m25eff1 = '';
  m25trm1 = '';
  m25memn = '';
  m25assc = '';
  m25pln2 = '';
  m25eff2 = '';
  m25trm2 = '';
  m25dob = '';
  m25pln3 = '';
  m25eff3 = '';
  m25trm3 = '';
  m25icd1 = '';
  m25icd2 = '';
  m25icd3 = '';
  m25xpa1 = '';
  m25xpa2 = '';
  m25atrm = '';
  m25err1 = '';
  m25err2 = '';
  m25err3 = '';
  screenChargeLines: ScreenChargeLine[] = [];
  screenExPay1 = '';
  screenExPay2 = '';
  screenAfterTerm = '';
}